+++ CHANGELOG.md
@@ -8,6 +8,10 @@
     * Moved `basic_complete` from utils into `cmd2.Cmd` class.
     * Moved `CompletionError` to exceptions.py
     * ``Namespace.__statement__`` has been removed. Use `Namespace.cmd2_statement.get()` instead.
+* Enhancements
+    * Added support for custom tab completion and up-arrow input history to `cmd2.Cmd2.read_input`.
+      See [read_input.py](https://github.com/python-cmd2/cmd2/blob/master/examples/read_input.py)
+      for an example.
 
 ## 1.4.0 (TBD, 2020)
 * Enhancements